+++ simple-schema-tests.js
@@ -514,7 +514,7 @@ Tinytest.add("SimpleSchema - Required Checks - Insert - Invalid", function(test)
   });
   // we should not get an error about optionalObject.requiredString because the whole object is null
   test.length(sc.invalidKeys(), 8);
-  
+
   sc = validateNoClean(ssr, {
     requiredString: null,
     requiredBoolean: null,
@@ -1141,7 +1141,7 @@ Tinytest.add("SimpleSchema - Type Checks - Insert", function(test) {
     string: ["test"]
   });
   test.length(sc2.invalidKeys(), 2); //without filter
-  
+
   sc = validate(ss, {
     string: ["test"]
   });
@@ -1188,7 +1188,7 @@ Tinytest.add("SimpleSchema - Type Checks - Insert", function(test) {
     boolean: ["test"]
   });
   test.length(sc2.invalidKeys(), 2); //without filter
-  
+
   sc = validate(ss, {
     boolean: ["test"]
   });
@@ -1230,7 +1230,7 @@ Tinytest.add("SimpleSchema - Type Checks - Insert", function(test) {
     number: ["test"]
   });
   test.length(sc2.invalidKeys(), 2); //without filter
-  
+
   sc = validate(ss, {
     number: ["test"]
   });
@@ -1284,7 +1284,7 @@ Tinytest.add("SimpleSchema - Type Checks - Insert", function(test) {
     date: ["test"]
   });
   test.length(sc2.invalidKeys(), 2); //without filter
-  
+
   sc = validate(ss, {
     date: ["test"]
   });
@@ -2560,49 +2560,46 @@ Tinytest.add("SimpleSchema - Allowed Values Checks - Update - Invalid - $set", f
 });
 
 Tinytest.add("SimpleSchema - Black Box Objects", function(test) {
-  var sc = validate(ss, {
+  function doTest(obj, isModifier, invalidCount) {
+    var sc = validate(ss, obj, isModifier, false, true);
+    test.length(sc.invalidKeys(), invalidCount);
+  }
+
+  doTest({
     blackBoxObject: "string"
-  }, false, false, true);
-  test.length(sc.invalidKeys(), 1);
+  }, false, 1);
 
-  var sc = validate(ss, {
+  doTest({
     blackBoxObject: {}
-  }, false, false, true);
-  test.length(sc.invalidKeys(), 0);
+  }, false, 0);
 
-  var sc = validate(ss, {
+  doTest({
     blackBoxObject: {
       foo: "bar"
     }
-  }, false, false, true);
-  test.length(sc.invalidKeys(), 0);
+  }, false, 0);
 
-  var sc = validate(ss, {$set: {
+  doTest({$set: {
     blackBoxObject: {
       foo: "bar"
     }
-  }}, true, false, true);
-  test.length(sc.invalidKeys(), 0);
+  }}, true, 0);
 
-  var sc = validate(ss, {$set: {
+  doTest({$set: {
     'blackBoxObject.foo': "bar"
-  }}, true, false, true);
-  test.length(sc.invalidKeys(), 0);
+  }}, true, 0);
 
-  var sc = validate(ss, {$set: {
+  doTest({$set: {
     'blackBoxObject.1': "bar"
-  }}, true, false, true);
-  test.length(sc.invalidKeys(), 0);
+  }}, true, 0);
 
-  var sc = validate(ss, {$push: {
+  doTest({$push: {
     'blackBoxObject.foo': "bar"
-  }}, true, false, true);
-  test.length(sc.invalidKeys(), 0);
+  }}, true, 0);
 
-  var sc = validate(ss, {$set: {
+  doTest({$set: {
     'blackBoxObject': []
-  }}, true, false, true);
-  test.length(sc.invalidKeys(), 1);
+  }}, true, 1);
 });
 
 Tinytest.add("SimpleSchema - Validation Against Another Key - Insert - Valid", function(test) {
@@ -2995,6 +2992,21 @@ Tinytest.add("SimpleSchema - Clean", function(test) {
       allowedStringsArray: {$each: ["tuna", "fish"]}
     }
   });
+
+  // Cleaning shouldn't remove anything within blackbox
+  doTest({blackBoxObject: {foo: 1}}, {blackBoxObject: {foo: 1}});
+  doTest({blackBoxObject: {foo: [1]}}, {blackBoxObject: {foo: [1]}});
+  doTest({blackBoxObject: {foo: [{bar: 1}]}}, {blackBoxObject: {foo: [{bar: 1}]}});
+  doTest({$set: {blackBoxObject: {foo: 1}}}, {$set: {blackBoxObject: {foo: 1}}});
+  doTest({$set: {blackBoxObject: {foo: [1]}}}, {$set: {blackBoxObject: {foo: [1]}}});
+  doTest({$set: {blackBoxObject: {foo: [{bar: 1}]}}}, {$set: {blackBoxObject: {foo: [{bar: 1}]}}});
+  doTest({$set: {'blackBoxObject.email.verificationTokens.$': {token: "Hi"}}}, {$set: {'blackBoxObject.email.verificationTokens.$': {token: "Hi"}}});
+  doTest({$set: {'blackBoxObject.email.verificationTokens.$.token': "Hi"}}, {$set: {'blackBoxObject.email.verificationTokens.$.token': "Hi"}});
+
+  doTest(
+    {$push: {'blackBoxObject.email.verificationTokens': {token: "Hi"}}},
+    {$push: {'blackBoxObject.email.verificationTokens': {token: "Hi"}}}
+    );
 
 });
 
@@ -3568,7 +3580,7 @@ Tinytest.add("SimpleSchema - AutoValues", function(test) {
     }
   });
   av2.clean({foo: "clown"});
-  
+
   var av3 = new SimpleSchema({
     foo: {
       type: String,
@@ -3593,7 +3605,7 @@ Tinytest.add("SimpleSchema - AutoValues", function(test) {
     }
   });
   av3.clean({foo: "clown", bar: true});
-  
+
   var av4 = new SimpleSchema({
     foo: {
       type: String,
@@ -3621,7 +3633,7 @@ Tinytest.add("SimpleSchema - AutoValues", function(test) {
   var doc = {bar: false};
   av4.clean(doc);
   test.equal(doc, {});
-  
+
   var av5 = new SimpleSchema({
     foo: {
       type: String,
@@ -3648,7 +3660,7 @@ Tinytest.add("SimpleSchema - AutoValues", function(test) {
   var doc = {$set: {bar: false}};
   av5.clean(doc);
   test.equal(doc, {$set: {bar: false}});
-  
+
   var av6 = new SimpleSchema({
     foo: {
       type: String,
@@ -3676,7 +3688,7 @@ Tinytest.add("SimpleSchema - AutoValues", function(test) {
   doc = {$set: {foo: "clown", bar: false}};
   av6.clean(doc);
   test.equal(doc, {$set: {foo: "clown", bar: true}});
-  
+
   var av7 = new SimpleSchema({
     foo: {
       type: String,
@@ -3846,6 +3858,14 @@ Tinytest.add("SimpleSchema - Optional Custom", function(test) {
   test.equal(ctx.invalidKeys()[0].type, 'custom', 'expected custom error');
 });
 
+Tinytest.add("SimpleSchema - Optional Conditional Requiredness With $unset", function(test) {
+  var ctx = optCust.namedContext();
+  // Ensure that custom requiredness works when string is $unset
+  ctx.validate({$unset: {"foo": ""}}, {modifier: true});
+  test.equal(ctx.invalidKeys().length, 1, 'expected 1 invalid key');
+  test.equal(ctx.invalidKeys()[0].type, 'custom', 'expected custom error');
+});
+
 Tinytest.add("SimpleSchema - Required Custom", function (test) {
   var ctx = reqCust.namedContext();
   // Ensure that we don't get required errors for a required field that
@@ -3915,6 +3935,7 @@ Tinytest.add("SimpleSchema - AllowsKey", function(test) {
   run('blackBoxObject.foo.$', true);
   run('blackBoxObject.foo.$foo', true);
   run('blackBoxObject.foo.$.$foo', true);
+  run('blackBoxObject.foo.bar.$.baz', true);
 });
 
 Tinytest.add("SimpleSchema - RegEx - Email", function (test) {